+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- [Added formatted gasLimit and gasPrice](https://github.com/multiversx/mx-sdk-dapp/pull/1444)
 - [Improve getPppuOptions readalibity](https://github.com/multiversx/mx-sdk-dapp/pull/1443)
 
 ## [[5.0.0-alpha.5](https://github.com/multiversx/mx-sdk-dapp/pull/1442)] - 2025-05-30